fix task
so... yes, it's done, praise our lord alberto!

I just need some fresh eyes, (and stop moving the pages, and maybe just maybe, stop doing thinks without thinking, so take a breath,)

I think the problem was that I got so many lines of codes, that I have lost my way.

Alberto help me with realise this, and we make it simple. (the new mantra)

we work on the Board view,  keep the focus on the data in taskStore.task, import such data from the useTaskStore

we also work on the edit and delete forms of the todotask, and it works fine,

still I need to make that the code autoactualize the new tasks, and also modify the done button!

lets keep working! YASSSS